test_generic_histogram_binning_calibration_by_feature or not
